Improve the performance of the custom classnames handling
