+++ packages/block-editor/src/hooks/custom-class-name.js
@@ -1,7 +1,6 @@
 /**
  * External dependencies
  */
-import { difference, omit } from 'lodash';
 import classnames from 'classnames';
 
 /**
@@ -10,11 +9,7 @@ import classnames from 'classnames';
 import { addFilter } from '@wordpress/hooks';
 import { TextControl } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
-import {
-	hasBlockSupport,
-	parseWithAttributeSchema,
-	getSaveContent,
-} from '@wordpress/blocks';
+import { hasBlockSupport } from '@wordpress/blocks';
 import { createHigherOrderComponent } from '@wordpress/compose';
 
 /**
@@ -117,63 +112,6 @@ export function addSaveProps( extraProps, blockType, attributes ) {
 	return extraProps;
 }
 
-/**
- * Given an HTML string, returns an array of class names assigned to the root
- * element in the markup.
- *
- * @param {string} innerHTML Markup string from which to extract classes.
- *
- * @return {string[]} Array of class names assigned to the root element.
- */
-export function getHTMLRootElementClasses( innerHTML ) {
-	innerHTML = `<div data-custom-class-name>${ innerHTML }</div>`;
-
-	const parsed = parseWithAttributeSchema( innerHTML, {
-		type: 'string',
-		source: 'attribute',
-		selector: '[data-custom-class-name] > *',
-		attribute: 'class',
-	} );
-
-	return parsed ? parsed.trim().split( /\s+/ ) : [];
-}
-
-/**
- * Given a parsed set of block attributes, if the block supports custom class
- * names and an unknown class (per the block's serialization behavior) is
- * found, the unknown classes are treated as custom classes. This prevents the
- * block from being considered as invalid.
- *
- * @param {Object} blockAttributes Original block attributes.
- * @param {Object} blockType       Block type settings.
- * @param {string} innerHTML       Original block markup.
- *
- * @return {Object} Filtered block attributes.
- */
-export function addParsedDifference( blockAttributes, blockType, innerHTML ) {
-	if ( hasBlockSupport( blockType, 'customClassName', true ) ) {
-		// To determine difference, serialize block given the known set of
-		// attributes, with the exception of `className`. This will determine
-		// the default set of classes. From there, any difference in innerHTML
-		// can be considered as custom classes.
-		const attributesSansClassName = omit( blockAttributes, [
-			'className',
-		] );
-		const serialized = getSaveContent( blockType, attributesSansClassName );
-		const defaultClasses = getHTMLRootElementClasses( serialized );
-		const actualClasses = getHTMLRootElementClasses( innerHTML );
-		const customClasses = difference( actualClasses, defaultClasses );
-
-		if ( customClasses.length ) {
-			blockAttributes.className = customClasses.join( ' ' );
-		} else if ( serialized ) {
-			delete blockAttributes.className;
-		}
-	}
-
-	return blockAttributes;
-}
-
 addFilter(
 	'blocks.registerBlockType',
 	'core/custom-class-name/attribute',
@@ -188,9 +126,4 @@ addFilter(
 	'blocks.getSaveContent.extraProps',
 	'core/custom-class-name/save-props',
 	addSaveProps
-);
-addFilter(
-	'blocks.getBlockAttributes',
-	'core/custom-class-name/addParsedDifference',
-	addParsedDifference
 );
