+++ packages/block-editor/src/hooks/custom-class-name.native.js
@@ -1,18 +1,13 @@
 /**
  * External dependencies
  */
-import { difference, compact } from 'lodash';
 import classnames from 'classnames';
 
 /**
  * WordPress dependencies
  */
 import { addFilter } from '@wordpress/hooks';
-import {
-	hasBlockSupport,
-	getSaveContent,
-	parseWithAttributeSchema,
-} from '@wordpress/blocks';
+import { hasBlockSupport } from '@wordpress/blocks';
 
 /**
  * Filters registered block settings, extending attributes with anchor using ID
@@ -61,64 +56,6 @@ export function addSaveProps( extraProps, blockType, attributes ) {
 	return extraProps;
 }
 
-/**
- * Given an HTML string, returns an array of class names assigned to the root
- * element in the markup.
- *
- * @param {string} innerHTML Markup string from which to extract classes.
- *
- * @return {string[]} Array of class names assigned to the root element.
- */
-export function getHTMLRootElementClasses( innerHTML ) {
-	innerHTML = `<div data-custom-class-name>${ innerHTML }</div>`;
-
-	const parsed = parseWithAttributeSchema( innerHTML, {
-		type: 'string',
-		source: 'attribute',
-		selector: '[data-custom-class-name] > *',
-		attribute: 'class',
-	} );
-
-	return parsed ? parsed.trim().split( /\s+/ ) : [];
-}
-
-/**
- * Given a parsed set of block attributes, if the block supports custom class
- * names and an unknown class (per the block's serialization behavior) is
- * found, the unknown classes are treated as custom classes. This prevents the
- * block from being considered as invalid.
- *
- * @param {Object} blockAttributes Original block attributes.
- * @param {Object} blockType       Block type settings.
- * @param {string} innerHTML       Original block markup.
- *
- * @return {Object} Filtered block attributes.
- */
-export function addParsedDifference( blockAttributes, blockType, innerHTML ) {
-	if ( hasBlockSupport( blockType, 'customClassName', true ) ) {
-		// To determine difference, serialize block given the known set of
-		// attributes. If there are classes which are mismatched with the
-		// incoming HTML of the block, add to filtered result.
-		const serialized = getSaveContent( blockType, blockAttributes );
-		const classes = getHTMLRootElementClasses( serialized );
-		const parsedClasses = getHTMLRootElementClasses( innerHTML );
-		const customClasses = difference( parsedClasses, classes );
-
-		const filteredClassName = compact( [
-			blockAttributes.className,
-			...customClasses,
-		] ).join( ' ' );
-
-		if ( filteredClassName ) {
-			blockAttributes.className = filteredClassName;
-		} else {
-			delete blockAttributes.className;
-		}
-	}
-
-	return blockAttributes;
-}
-
 addFilter(
 	'blocks.registerBlockType',
 	'core/custom-class-name/attribute',
@@ -128,9 +65,4 @@ addFilter(
 	'blocks.getSaveContent.extraProps',
 	'core/custom-class-name/save-props',
 	addSaveProps
-);
-addFilter(
-	'blocks.getBlockAttributes',
-	'core/custom-class-name/addParsedDifference',
-	addParsedDifference
 );
