+++ packages/blocks/src/api/validation/index.js
@@ -13,6 +13,7 @@ import { decodeEntities } from '@wordpress/html-entities';
  * Internal dependencies
  */
 import { createLogger, createQueuedLogger } from './logger';
+import { fixCustomClassname } from './fix-custom-classname';
 import { getSaveContent } from '../serializer';
 import { normalizeBlockType } from '../utils';
 
@@ -697,6 +698,7 @@ export function getBlockContentValidationResult(
 ) {
 	const blockType = normalizeBlockType( blockTypeOrName );
 	let generatedBlockContent;
+	let fixedBlockAttributes = attributes;
 	try {
 		generatedBlockContent = getSaveContent( blockType, attributes );
 	} catch ( error ) {
@@ -711,11 +713,31 @@ export function getBlockContentValidationResult(
 		};
 	}
 
-	const isValid = isEquivalentHTML(
+	let isValid = isEquivalentHTML(
 		originalBlockContent,
 		generatedBlockContent,
 		logger
 	);
+
+	// Try fixing the custom classnames
+	if ( ! isValid ) {
+		fixedBlockAttributes = fixCustomClassname(
+			attributes,
+			blockType,
+			originalBlockContent
+		);
+		generatedBlockContent = getSaveContent(
+			blockType,
+			fixedBlockAttributes
+		);
+
+		isValid = isEquivalentHTML(
+			originalBlockContent,
+			generatedBlockContent,
+			logger
+		);
+	}
+
 	if ( ! isValid ) {
 		logger.error(
 			'Block validation failed for `%s` (%o).\n\nContent generated by `save` function:\n\n%s\n\nContent retrieved from post body:\n\n%s',
@@ -729,6 +751,7 @@ export function getBlockContentValidationResult(
 	return {
 		isValid,
 		validationIssues: logger.getItems(),
+		attributes: fixedBlockAttributes,
 	};
 }
